(23/05) Início de implementação de segurança de API no back-end

+++ back-end/src/controllers/users.js
@@ -14,6 +14,11 @@ const controller = {}     // Objeto vazio
 
 controller.create = async function (req, res) {
   try {
+
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if (!req?.authUser?.is_admin) return res.status(403).end()
+
     // Caso exista o campo "password" em req.body, é
     // necessário gerar o hash da senha antes de
     // armazená-la no BD, usando o algoritmo argon2
@@ -39,8 +44,14 @@ controller.create = async function (req, res) {
 
 controller.retrieveAll = async function (req, res) {
   try {
-    const result = await prisma.user.findMany()
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden
+    if (!req?.authUser?.is_admin) return res.status(403).end()
 
+    // const result = await prisma.user.findMany()
+    const result = await prisma.user.findMany({
+      omit: { password: true }
+    })
     // HTTP 200: OK (implícito)
     res.send(result)
   }
@@ -54,9 +65,18 @@ controller.retrieveAll = async function (req, res) {
 
 controller.retrieveOne = async function (req, res) {
   try {
+    // Somente usuários administradores ou o próprio usuário
+    // autenticado podem acessar este recurso
+    // HTTP 403: Forbidden
+    if (!(req?.authUser?.is_admin ||
+      Number(req?.authUser?.id) === Number(req.params.id)))
+      return res.status(403).end()
+
     const result = await prisma.user.findUnique({
+      omit: { password: true },
       where: { id: Number(req.params.id) }
     })
+
 
     // Encontrou ~> retorna HTTP 200: OK (implícito)
     if (result) res.send(result)
@@ -73,6 +93,14 @@ controller.retrieveOne = async function (req, res) {
 
 controller.update = async function (req, res) {
   try {
+
+    // Somente usuários administradores ou o próprio usuário
+    // autenticado podem acessar este recurso
+    // HTTP 403: Forbidden
+    if (!(req?.authUser?.is_admin ||
+      Number(req?.authUser?.id) === Number(req.params.id)))
+      return res.status(403).end()
+
     // Caso exista o campo "password" em req.body, é
     // necessário gerar o hash da senha antes de
     // armazená-la no BD, usando o algoritmo argon2
@@ -103,6 +131,11 @@ controller.update = async function (req, res) {
 
 controller.delete = async function (req, res) {
   try {
+
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden
+    if (!req?.authUser?.is_admin) return res.status(403).end()
+
     await prisma.user.delete({
       where: { id: Number(req.params.id) }
     })
